pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
dynamic = ["version"]
name = "mpay"
description = "Track financial transactions in a home environment"
readme = "README.md"
authors = [
    {name = "Ondřej Sluka"},
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "PyYAML >= 6.0",
    "SQLAlchemy >= 2.0.35",
    "pandas >= 2.2.3",
]


[project.urls]
Homepage = "https://github.com/ondras12345/mpay"
Repository = "https://github.com/ondras12345/mpay"
Issues = "https://github.com/ondras12345/mpay/issues"


[project.scripts]
mpay = "mpay:main"
